+++ client/src/pages/Friend/LetterItem/LetterItemStyle.jsx
@@ -1,17 +1,6 @@
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
 
-const LetterWrapper = styled.ul`
-  overflow: hidden;
-  position: relative;
-  margin: 15rem 1rem 20rem 1rem;
-  display: flex;
-  flex-wrap: wrap;
-  width: 90%;
-  gap: 1.5rem;
-  text-decoration: none;
-`;
-
 const Letter = styled.li`
   width: 300px;
   border-radius: 1rem;
@@ -105,8 +94,8 @@ const WriteBtn = styled.button`
     color: #ccc;
   }
 `;
+
 export {
-  LetterWrapper,
   Letter,
   StyledLink,
   LetterHeader,
